catalog: more data from ROME 4.0
The API allows to select a list of fields, to get more data. This is
great so now we have access to profession's descriptions.

Sadly, within the self-imposed time constraint, finding a way to select
"complex field" is not achievable. As a result, I have to  let go of
"formacodes". For now. This problem can probably be solved by sending an
email to France Travail's support team to see what I missed from the
documentation.

Anyway, minor updates are required to both rome_extract and
rome_transform, and the new data is displayed on the detailed view.